+++ lib/playwright_ex/processes/connection.ex
@@ -11,7 +11,7 @@ defmodule PlaywrightEx.Connection do
 
   import Kernel, except: [send: 2]
 
-  alias PlaywrightEx.FrameEventRecorder
+  alias PlaywrightEx.Resource.Frame, as: FrameResource
 
   @timeout_grace_factor 1.5
   @min_genserver_timeout to_timeout(second: 1)
@@ -108,7 +108,7 @@ defmodule PlaywrightEx.Connection do
 
   @doc false
   def pending(:cast, {:playwright_msg, %{method: :__create__, params: %{guid: "Playwright"}} = msg}, data) do
-    {:next_state, :started, handle_create(data, msg)}
+    {:next_state, :started, cache_initializer(data, msg)}
   end
 
   def pending(:cast, _msg, _data), do: {:keep_state_and_data, [:postpone]}
@@ -171,38 +171,28 @@ defmodule PlaywrightEx.Connection do
   end
 
   def started(:cast, {:playwright_msg, msg}, data) do
-    {:keep_state,
-     data |> handle_create(msg) |> maybe_start_frame_event_recorder(msg) |> notify_subscribers(msg) |> handle_dispose(msg)}
+    data = cache_initializer(data, msg)
+    _ = FrameResource.maybe_start(data.config.name, msg)
+    data = notify_subscribers(data, msg)
+
+    {:keep_state, release_disposed_guid(data, msg)}
   end
 
-  defp handle_create(data, %{method: :__create__} = msg) do
+  defp cache_initializer(data, %{method: :__create__} = msg) do
     put_in(data.initializers[msg.params.guid], msg.params.initializer)
   end
 
-  defp handle_create(data, _msg), do: data
+  defp cache_initializer(data, _msg), do: data
 
-  defp maybe_start_frame_event_recorder(data, %{
-         method: :__create__,
-         params: %{guid: guid, initializer: %{url: _url, load_states: _load_states} = initializer}
-       }) do
-    _ = FrameEventRecorder.ensure_started(data.config.name, guid, initializer)
-    data
-  end
+  defp release_disposed_guid(data, %{method: :__dispose__} = msg) do
+    _ = FrameResource.maybe_stop(data.config.name, msg.guid)
 
-  defp maybe_start_frame_event_recorder(data, %{method: :__create__}) do
-    data
-  end
-
-  defp maybe_start_frame_event_recorder(data, _msg), do: data
-
-  defp handle_dispose(data, %{method: :__dispose__} = msg) do
     data
     |> Map.update!(:initializers, &Map.delete(&1, msg.guid))
-    |> stop_disposed_frame_event_recorder(msg.guid)
     |> clear_disposed_guid_subscribers(msg.guid)
   end
 
-  defp handle_dispose(data, _msg), do: data
+  defp release_disposed_guid(data, _msg), do: data
 
   defp notify_subscribers(data, %{guid: guid} = msg) do
     for pid <- :pg.get_members(data.config.pg_scope, pg_group(guid)) do
@@ -223,11 +213,6 @@ defmodule PlaywrightEx.Connection do
       _ = :pg.leave(data.config.pg_scope, group, pid)
     end
 
-    data
-  end
-
-  defp stop_disposed_frame_event_recorder(data, guid) do
-    _ = FrameEventRecorder.terminate_frame(data.config.name, guid)
     data
   end